+++ paper/src/main/java/dev/qixils/crowdcontrol/plugin/paper/commands/LootboxCommand.java
@@ -22,16 +22,28 @@ import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 @Getter
 public class LootboxCommand extends ImmediateCommand {
+	private static final List<Material> ITEMS = Arrays.stream(Material.values()).filter(Material::isItem).toList();
+	private static final Set<Material> GOOD_ITEMS = ITEMS.stream().filter(material ->
+			material.getMaxDurability() > 1
+					|| material == Material.GOLDEN_APPLE
+					|| material == Material.ENCHANTED_GOLDEN_APPLE
+					|| material == Material.DIAMOND_BLOCK
+					|| material == Material.NETHERITE_BLOCK
+					|| material == Material.IRON_BLOCK
+					|| material == Material.GOLD_BLOCK
+	).collect(Collectors.toUnmodifiableSet());
+	private static final List<Attribute> ATTRIBUTES = Arrays.asList(Attribute.values());
 	private final String effectName = "lootbox";
 	private final String displayName = "Open Lootbox";
 
@@ -39,72 +51,124 @@ public class LootboxCommand extends ImmediateCommand {
 		super(plugin);
 	}
 
+	private static boolean isGoodItem(@Nullable Material item) {
+		return item != null && GOOD_ITEMS.contains(item);
+	}
+
+	/**
+	 * Creates a random item that is influenced by the supplied luck value.
+	 * The item may contain enchantments and modifiers if applicable.
+	 *
+	 * @param luck zero-indexed level of luck
+	 * @return new random item
+	 */
+	public static ItemStack createRandomItem(int luck) {
+		// determine the item used in the stack
+		// "good" items have a higher likelihood of being picked with positive luck
+		List<Material> items = new ArrayList<>(ITEMS);
+		Collections.shuffle(items, random);
+		Material item = null;
+		for (int i = 0; i <= luck * 3; i++) {
+			Material oldItem = item;
+			item = items.get(i);
+			if (isGoodItem(item) && !isGoodItem(oldItem))
+				break;
+		}
+		assert item != null;
+
+		// determine the size of the item stack
+		int quantity = 1;
+		if (item.getMaxStackSize() > 1) {
+			for (int i = 0; i <= luck; i++) {
+				quantity = Math.max(quantity, RandomUtil.nextInclusiveInt(1, item.getMaxStackSize()));
+			}
+		}
+
+		// create item stack
+		ItemStack itemStack = new ItemStack(item, quantity);
+		ItemMeta itemMeta = itemStack.getItemMeta();
+		// make item unbreakable with a default chance of 10% (up to 100% at 6 luck)
+		if (random.nextDouble() >= (0.9D - (luck * .15D)))
+			itemMeta.setUnbreakable(true);
+
+		// determine enchantments to add
+		int enchantments = 0;
+		for (int i = 0; i <= luck; i++) {
+			enchantments = Math.max(enchantments, RandomUtil.weightedRandom(EnchantmentWeights.values(), EnchantmentWeights.TOTAL_WEIGHTS).getLevel());
+		}
+		List<Enchantment> enchantmentList = Arrays.stream(EnchantmentWrapper.values())
+				.filter(enchantment -> enchantment.canEnchantItem(itemStack))
+				.collect(Collectors.toList());
+		// add enchantments
+		if (enchantments > 0 && !enchantmentList.isEmpty()) {
+			Collections.shuffle(enchantmentList, random);
+			List<Enchantment> addedEnchantments = new ArrayList<>(enchantments);
+			int count = 0;
+			for (int i = 0; i < enchantmentList.size() && count < enchantments; ++i) {
+				Enchantment enchantment = enchantmentList.get(i);
+				// block conflicting enchantments (unless the die roll decides otherwise)
+				if (addedEnchantments.stream().anyMatch(x -> x.conflictsWith(enchantment)) && random.nextDouble() >= (.1d + (luck * .1d)))
+					continue;
+				++count;
+				addedEnchantments.add(enchantment);
+				// determine enchantment level
+				int level = enchantment.getStartLevel();
+				if (enchantment.getMaxLevel() > enchantment.getStartLevel()) {
+					for (int j = 0; j <= luck; j++) {
+						level = Math.max(level, RandomUtil.nextInclusiveInt(enchantment.getStartLevel(), enchantment.getMaxLevel()));
+					}
+					if (random.nextDouble() >= (0.5D - (luck * .07D)))
+						level += random.nextInt(4);
+				}
+				// add enchant
+				itemStack.addEnchantment(enchantment, level);
+			}
+		}
+
+		// determine attributes to add
+		int attributes = 0;
+		for (int i = 0; i <= luck; i++) {
+			attributes = Math.max(attributes, RandomUtil.weightedRandom(AttributeWeights.values(), AttributeWeights.TOTAL_WEIGHTS).getLevel());
+		}
+		// add attributes
+		if (attributes > 0) {
+			List<Attribute> attributeList = new ArrayList<>(ATTRIBUTES);
+			Collections.shuffle(attributeList, random);
+			for (int i = 0; i < attributeList.size() && i < attributes; ++i) {
+				Attribute attribute = attributeList.get(i);
+				String name = "lootbox_" + attribute.getKey().getKey();
+				// determine percent amount for the modifier
+				double amount = 0d;
+				for (int j = 0; j <= luck; j++) {
+					amount = Math.max(amount, (random.nextDouble() * 2) - 1);
+				}
+				// create & add attribute
+				AttributeModifier attributeModifier = new AttributeModifier(name, amount, AttributeModifier.Operation.MULTIPLY_SCALAR_1);
+				itemMeta.addAttributeModifier(attribute, attributeModifier);
+			}
+		}
+
+		// finish up
+		itemStack.setItemMeta(itemMeta);
+		return itemStack;
+	}
+
 	@Override
 	public Response.@NotNull Builder executeImmediately(@NotNull List<@NotNull Player> players, @NotNull Request request) {
 		for (Player player : players) {
 			Inventory lootbox = Bukkit.createInventory(null, 27, CommandConstants.buildLootboxTitle(request));
-			List<Material> items = new ArrayList<>(Arrays.asList(Material.values()));
-			Collections.shuffle(items, random);
-			Material item = null;
-			for (Material i : items) {
-				if (i.isItem()) {
-					item = i;
-					break;
-				}
-			}
-			assert item != null;
-
-			ItemStack itemStack = new ItemStack(item, 1 + random.nextInt(item.getMaxStackSize()));
-			ItemMeta itemMeta = itemStack.getItemMeta();
+			// create item
+			ItemStack randomItem = createRandomItem(0);
+			ItemMeta itemMeta = randomItem.getItemMeta();
 			itemMeta.lore(Collections.singletonList(CommandConstants.buildLootboxLore(request)));
-			if (random.nextDouble() >= 0.9D) {
-				itemMeta.setUnbreakable(true);
-			}
-
-			// big dumb enchantment logic to generate sane items lmfao
-			int enchantments = RandomUtil.weightedRandom(EnchantmentWeights.values(), EnchantmentWeights.TOTAL_WEIGHTS).getLevel();
-			if (enchantments > 0) {
-				List<Enchantment> enchantmentList = new ArrayList<>();
-				for (Enchantment enchantment : EnchantmentWrapper.values()) {
-					if (enchantment.canEnchantItem(itemStack)) enchantmentList.add(enchantment);
-				}
-				if (!enchantmentList.isEmpty()) {
-					Collections.shuffle(enchantmentList, random);
-					Set<Enchantment> addedEnchantments = new HashSet<>();
-					int count = 0;
-					for (int i = 0; i < enchantmentList.size() && count < enchantments; ++i) {
-						Enchantment enchantment = enchantmentList.get(i);
-						if (addedEnchantments.stream().noneMatch(x -> x.conflictsWith(enchantment))) {
-							++count;
-							addedEnchantments.add(enchantment);
-							int level = enchantment.getStartLevel() + random.nextInt(enchantment.getMaxLevel() - enchantment.getStartLevel() + 1);
-							if (random.nextBoolean()) // bonus OP enchant chance
-								level += random.nextInt(4);
-							itemStack.addEnchantment(enchantment, level);
-						}
-					}
-				}
-			}
-
-			int attributes = RandomUtil.weightedRandom(AttributeWeights.values(), AttributeWeights.TOTAL_WEIGHTS).getLevel();
-			if (attributes > 0) {
-				List<Attribute> attributeList = Arrays.asList(Attribute.values());
-				Collections.shuffle(attributeList, random);
-				for (int i = 0; i < attributeList.size() && i < attributes; ++i) {
-					Attribute attribute = attributeList.get(i);
-					String name = "lootbox_" + attribute.getKey().getKey();
-					AttributeModifier attributeModifier = new AttributeModifier(name, (random.nextDouble() * 2) - 1, AttributeModifier.Operation.MULTIPLY_SCALAR_1);
-					itemMeta.addAttributeModifier(attribute, attributeModifier);
-				}
-			}
-
-			itemStack.setItemMeta(itemMeta);
-			lootbox.setItem(13, itemStack);
+			randomItem.setItemMeta(itemMeta);
+			// display lootbox
+			lootbox.setItem(13, randomItem);
+			sync(() -> player.openInventory(lootbox));
 			player.playSound(
 					Sounds.LOOTBOX_CHIME.get(),
 					Sound.Emitter.self()
 			);
-			sync(() -> player.openInventory(lootbox));
 		}
 		return request.buildResponse().type(Response.ResultType.SUCCESS);
 	}