+++ paper/src/main/java/dev/qixils/crowdcontrol/plugin/paper/commands/executeorperish/StandOnBlockCondition.java
@@ -12,19 +12,21 @@ import java.util.Set;
 
 @Getter
 public class StandOnBlockCondition implements SuccessCondition {
+	private final int rewardLuck;
 	private final Set<Material> blocks;
 	private final Component component;
 
-	public StandOnBlockCondition(String displayText, Material displayItem, Material... otherItems) {
+	public StandOnBlockCondition(int rewardLuck, String displayText, Material displayItem, Material... otherItems) {
+		this.rewardLuck = rewardLuck;
 		this.blocks = EnumSet.of(displayItem, otherItems);
-
-		component = Component.text("Stand on ").append(Component.text(displayText)
+		this.component = Component.text("Stand on ").append(Component.text(displayText)
 				.replaceText(builder -> builder.matchLiteral("%s").once()
 						.replacement(Component.translatable(displayItem).color(NamedTextColor.GREEN)))
 		);
 	}
 
-	public StandOnBlockCondition(Component display, Material first, Material... other) {
+	public StandOnBlockCondition(int rewardLuck, Component display, Material first, Material... other) {
+		this.rewardLuck = rewardLuck;
 		this.blocks = EnumSet.of(first, other);
 		this.component = Component.text("Stand on ").append(display);
 	}